reformat testing files
