+++ test/CreatePropertyController.test.ts
@@ -1,4 +1,5 @@
-import { createProperty, deleteProperty } from "../src/controllers/PropertyController"
+import { createProperty } from "../src/controllers/CreatePropertyController"
+import { deleteProperty } from "../src/controllers/DeletePropertyController"
 
 const bare_bones_property = {
     address: "123-testing-st",
@@ -23,16 +24,4 @@ test("create-property", async () => {
         // always attempt to delete the created property
         await deleteProperty(property.property_id, property.image_url)
     }
-})
-
-test("delete-property", async () => {
-    const id = -2;
-    const property = {
-        ...bare_bones_property,
-        property_id: id,
-    }
-
-    await createProperty(property, test_rooms);
-    let res = await deleteProperty(property.property_id, property.image_url)
-    expect(res).toBe(id)
 })
